+++ src/main/java/insane96mcp/survivalreimagined/module/experience/enchanting/EnsorcellerBlock.java
@@ -1,14 +1,19 @@
 package insane96mcp.survivalreimagined.module.experience.enchanting;
 
 import net.minecraft.core.BlockPos;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.Containers;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.MenuProvider;
 import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.BaseEntityBlock;
@@ -21,6 +26,8 @@ import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
 import org.jetbrains.annotations.Nullable;
+
+import java.util.List;
 
 public class EnsorcellerBlock extends BaseEntityBlock {
 
@@ -69,11 +76,30 @@ public class EnsorcellerBlock extends BaseEntityBlock {
                 }
 
                 pLevel.updateNeighbourForOutputSignal(pPos, this);
-                ensorcellerBlockEntity.dropExperience(pLevel);
+                //ensorcellerBlockEntity.dropExperience(pLevel);
             }
 
             super.onRemove(pState, pLevel, pPos, pNewState, pIsMoving);
         }
+    }
+
+    @Override
+    public void playerWillDestroy(Level level, BlockPos pos, BlockState state, Player player) {
+        BlockEntity blockentity = level.getBlockEntity(pos);
+        if (blockentity instanceof EnsorcellerBlockEntity ensorcellerBlockEntity) {
+            if (!level.isClientSide) {
+                ItemStack itemstack = new ItemStack(this.asItem());
+                blockentity.saveToItem(itemstack);
+                if (ensorcellerBlockEntity.hasCustomName()) {
+                    itemstack.setHoverName(ensorcellerBlockEntity.getCustomName());
+                }
+
+                ItemEntity itementity = new ItemEntity(level, (double)pos.getX() + 0.5D, (double)pos.getY() + 0.5D, (double)pos.getZ() + 0.5D, itemstack);
+                itementity.setDefaultPickUpDelay();
+                level.addFreshEntity(itementity);
+            }
+        }
+        super.playerWillDestroy(level, pos, state, player);
     }
 
     /**
@@ -91,5 +117,14 @@ public class EnsorcellerBlock extends BaseEntityBlock {
 
     public boolean isPathfindable(BlockState pState, BlockGetter pLevel, BlockPos pPos, PathComputationType pType) {
         return false;
+    }
+
+    @Override
+    public void appendHoverText(ItemStack stack, @Nullable BlockGetter pLevel, List<Component> tooltip, TooltipFlag pFlag) {
+        CompoundTag tag = BlockItem.getBlockEntityData(stack);
+        int steps = 0;
+        if (tag != null)
+            steps = tag.getInt("Steps");
+        tooltip.add(Component.literal("Levels rolled: %d".formatted(steps)));
     }
 }