+++ changelog.md
@@ -25,13 +25,13 @@
   * Is no longer rendered way up in the air and down to the bottom of the world 
   * Is now 60% transparent
 * Increased flint base damage by 0.5
+* Ensorceller now keeps levels rolled when broken
 * Mushrooms are now bone meal-able only when on mycelium
 * Added Melee Protection enchantment
 * Arrows can now be made from cloth in a fletching table
 * Slight increase to Unbreaking formula (x1.29, x1.67, x2.11, x2.5 -> x1.33, x1.82, x2.5, x3.33 durability at levels I to IV)
 * Slightly decreased Durium Ore XP drop
 * Blasting is now affected by Water, midair, haste and mining fatigue
-* Breaking an ensorceller now gives back some experience used to roll
 * Pillagers now drop emeralds and arrows
 * Phantoms now actually spawn in The End
 * Fixed data packs not disabling if they were already enabled in a world